+++ src/lib.rs
@@ -6,7 +6,9 @@
 
 pub use bit_string::BitString;
 pub use bit_vector::{BitVector, BitVectorBuilder};
+pub use louds::{Louds, LoudsBuilder, LoudsIndex, LoudsNodeNum};
 
 pub mod bit_string;
 pub mod bit_vector;
 mod internal_data_structure;
+pub mod louds;